requirements.txt
fastapi~=0.115.11
pydantic~=2.10.6
uvicorn
transformers
torch
python-dotenv
PyYAML~=6.0.2
accelerate